requirements.txt
Django>=5.1.5
djangorestframework>=3.13.1
django-ckeditor>=6.0
googletrans==4.0.0-rc1
django-redis>=5.0.0
